requirements.txt
llama-index-llms-ollama
llama-index-embeddings-huggingface